Change dark mode background color
When using the dark mode in high nit devices, it became really apparent
the fact that the background was really blue, making the website ugly in
general. To combat this, the background color was changed to a monotone
blackish color, as to not draw more attention to the background than to
the foreground (the content of the website).