Change anchor style
While creating the résumé page I noticed that the user could confuse
normal highlights with link highlights, since they were basically the
same. To remedy this I added an arrow symbol at the end of every link, to
signify that if the user clicks it he will be redirected to another
place, while normal highlights continue as is.

Note that I did not add arrows to the footer. Since, in my opinion, it
became really ugly with the amount of arrows present. Although, it should
be clear that the highlights in the footer belong to clickable links
only.