Major refactor (again)
This time I decided to change the layout and colors of the website, I'm
still not 100% satisfied with the end result but it's better than
before. There were a lot of changes that I'm not really going to list
because this is a solo project and I know what I did in the end.