Fix some complaints
Firstly, lighthouse keeps nagging me about text visibility when I use the
`--color-inactive` color, so I changed it to a darker color (even though
it was prettier before, in my opinion).

Secondly, The `<article>` tag started with an `<h3>` in the about page,
so I corrected it changing it to a `<h1>`.